Issue 16: Delete tables
