--- backupper-gui/src/main/java/pl/gda/pg/eti/autyzm/Model/Table.java
@@ -1,8 +1,0 @@
-package pl.gda.pg.eti.autyzm.Model;
-
-
-public abstract class Table {
-
-    abstract void setDataBindings();
-    abstract void setColumnWidth();
-}
