+++ backupper-gui/src/main/resources/main.css
@@ -3,46 +3,35 @@
 }
 
 /*Download copy tab*/
-.copyTab{
+.backupTab{
     -fx-border-style: none;
     -fx-spacing: 10px;
     -fx-padding: 30px;
 }
 
-.copyNameInputArea {
+.backupNameInputArea {
     -fx-margin: 30px ;
     -fx-padding: 10px;
 }
 
-.chooseDeviceLabel{
-    -fx-font-style: italic;
-}
+/*Refresh copy tab*/
 
-.chooseDeviceArea{
+.deviceArea{
     -fx-margin: 30px ;
     -fx-padding: 10px;
     -fx-spacing: 10px;
     -fx-alignment: center;
 }
 
-.downloadCopyButtonArea {
-    -fx-alignment: center;
-}
 
-.downloadCopyButton {
-    -fx-min-width: 350px;
-    -fx-alignment: center;
-}
 
-/*Refresh copy tab*/
-
-.refreshTab{
+.restoreTab{
     -fx-border-style: none;
     -fx-spacing: 10px;
     -fx-margin: 30px ;
     -fx-padding: 10px;
 }
 
-.refreshActionColumn{
+.restoreActionColumn{
     -fx-alignment: center;
 }
